+++ main/main.go
@@ -2,7 +2,10 @@ package main
 
 import (
 	"fmt"
+	"hapemu/accessor"
+	"hapemu/cosine"
 	"hapemu/service"
+	"hapemu/vector"
 	"log"
 	"net/http"
 
@@ -10,11 +13,15 @@ import (
 )
 
 func main() {
+	cosineSimilarityService := cosine.NewCosineSimilarityService()
+	vectorGeneratorService := vector.NewCosineSimilarityService()
+	hapemuDatabaseAccessor := accessor.NewHapemuDatabaseAccessor()
+	recommendationService := service.NewRecommendationService(cosineSimilarityService, vectorGeneratorService, hapemuDatabaseAccessor)
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "Hello World")
 	})
 	http.HandleFunc("/send-email", service.EmailRecommendations)
-	http.HandleFunc("/get-recommendations", service.RecommendSmartphones)
+	http.HandleFunc("/get-recommendations", recommendationService.RecommendSmartphones)
 
 	handler := cors.Default().Handler(http.DefaultServeMux)
